requirements/base.txt
# requirements/base.txt
# BI Platform Dependencies
# Using flexible version constraints for better Python 3.12 compatibility
pandas>=1.5.3,<3.0.0
plotly>=5.10.0,<6.0.0
dash>=2.7.1,<3.0.0
dash-bootstrap-components>=1.3.0,<2.0.0
dash-table>=5.0.0
sqlalchemy>=1.4.46,<3.0.0
openpyxl>=3.0.10,<4.0.0

# API Engine Dependencies  
requests>=2.28.2,<3.0.0
python-dotenv>=0.19.0,<2.0.0
pyyaml>=6.0.1,<7.0.0
boto3>=1.26.0,<2.0.0
python-decouple>=3.8,<4.0.0

# Common Utilities
click>=8.1.3,<9.0.0
loguru>=0.6.0,<1.0.0
pytest>=7.2.0,<8.0.0

# HTTP Service
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0

# Chart Export
kaleido>=0.2.1,<1.0.0
reportlab>=4.0.7,<5.0.0

# Note: Database drivers (psycopg2-binary, pymysql) are optional
# Install from requirements/database.txt if needed
